renamed NewCreateLook to CreateLookTabs
